+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -12,11 +13,11 @@
         android:theme="@style/Theme.MovieSearch"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="view.MainActivity"
             android:exported="false"
             android:theme="@style/Theme.MovieSearch.NoActionBar" />
         <activity
-            android:name=".SplashScreenActivity"
+            android:name="view.SplashScreenActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
