+++ app/src/main/java/domain/Film.kt
@@ -1,7 +1,8 @@
-package com.example.moviesearch
+package domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+
 @Parcelize
 data class Film(
     val title: String,
