+++ app/src/main/java/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package domain
+
+import data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
