+++ app/src/main/java/utils/AnimationHelper.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearch
+package utils
 
 import android.app.Activity
 import android.view.View
