+++ app/src/main/java/view/SplashScreenActivity.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearch
+package view
 
 import android.content.Intent
 import android.os.Bundle
