+++ app/src/main/java/view/fragments/SelectionsFragment.kt
@@ -1,37 +1,34 @@
-package com.example.moviesearch
+package view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.moviesearch.databinding.FragmentWatchLaterBinding
+import com.example.moviesearch.databinding.FragmentSelectionsBinding
+import utils.AnimationHelper
 
-class WatchLaterFragment : Fragment() {
-    private var _binding: FragmentWatchLaterBinding? = null
-    private val binding: FragmentWatchLaterBinding
+
+class SelectionsFragment : Fragment() {
+    private var _binding: FragmentSelectionsBinding? = null
+    private val binding: FragmentSelectionsBinding
         get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentWatchLaterBinding.inflate(layoutInflater)
+        _binding = FragmentSelectionsBinding.inflate(inflater, container, false)
         return binding.root
-
-
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
         AnimationHelper.performFragmentCircularRevealAnimation(
-            binding.fragmentWatchLater,
+            binding.fragmeentSelections,
             requireActivity(),
-            3
+            4
         )
-
     }
-
 
 }
