+++ app/src/main/java/view/fragments/WatchLaterFragment.kt
@@ -1,33 +1,38 @@
-package com.example.moviesearch
+package view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.moviesearch.databinding.FragmentSelectionsBinding
+import com.example.moviesearch.databinding.FragmentWatchLaterBinding
+import utils.AnimationHelper
 
-
-class SelectionsFragment : Fragment() {
-    private var _binding: FragmentSelectionsBinding? = null
-    private val binding: FragmentSelectionsBinding
+class WatchLaterFragment : Fragment() {
+    private var _binding: FragmentWatchLaterBinding? = null
+    private val binding: FragmentWatchLaterBinding
         get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSelectionsBinding.inflate(inflater, container, false)
+        _binding = FragmentWatchLaterBinding.inflate(layoutInflater)
         return binding.root
+
+
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
         AnimationHelper.performFragmentCircularRevealAnimation(
-            binding.fragmeentSelections,
+            binding.fragmentWatchLater,
             requireActivity(),
-            4
+            3
         )
+
     }
+
 
 }
