+++ app/src/main/java/view/rv_viewholders/FilmViewHolder.kt
@@ -1,8 +1,9 @@
-package com.example.moviesearch
+package view.rv_viewholders
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.moviesearch.databinding.FilmItemBinding
+import domain.Film
 
 class FilmViewHolder(binding: FilmItemBinding) : RecyclerView.ViewHolder(binding.root) {
     private val title = binding.title
