+++ app/src/main/java/viewmodel/FavoriteFragmentViewModel.kt
@@ -1,0 +1,17 @@
+package viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.example.moviesearch.App
+import domain.Film
+import domain.Interactor
+
+class FavoriteFragmentViewModel: ViewModel() {
+    val filmsListLiveData: MutableLiveData<List<Film>> = MutableLiveData()
+    private var interactor: Interactor = App.instance.interactor
+
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
