+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/shape_background_main_second"
-    tools:context=".MainActivity">
+    tools:context="view.MainActivity">
 
 
     <FrameLayout
