+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/shape_background_main_second"
-    tools:context=".SplashScreenActivity">
+    tools:context="view.SplashScreenActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:layout_width="350dp"
