+++ app/src/main/res/layout/film_item.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="150dp">
 
-    <com.example.moviesearch.RatingDonutView
+    <view.customeviews.RatingDonutView
         android:id="@+id/rating_donut"
         android:layout_width="45dp"
         android:layout_height="45dp"
