+++ app/src/main/res/layout/fragment_favorites.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/shape_background_favorites"
     android:visibility="invisible"
     tools:visibility="visible"
-    tools:context=".FavoritesFragment">
+    tools:context="view.fragments.FavoritesFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
