+++ app/src/main/res/layout/fragment_selections.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/shape_background_selections"
     android:visibility="invisible"
-    tools:context=".SelectionsFragment"
+    tools:context="view.fragments.SelectionsFragment"
     tools:visibility="visible">
