+++ app/src/main/res/values-ru/strings.xml
@@ -17,6 +17,7 @@
     <string name="home">Главная</string>
     <string name="toast_home">Главная</string>
     <string name="seach_view_hint">введите название фильма</string>
+    <string name="toast_favorites">Избранное</string>
 
 
 </resources>
