+++ app/src/main/res/values/strings.xml
@@ -16,6 +16,7 @@
     <string name="home">Home</string>
     <string name="toast_home">Home</string>
     <string name="seach_view_hint">enter film title</string>
+    <string name="toast_favorites">Favorites</string>
 
 
 </resources>
